+++ src/Providers/Mail/Mailer.php
@@ -44,13 +44,6 @@ class Mailer extends MailerAbstract {
 	 */
 	public function is_mailer_complete() {
 
-		// Do not allow to send test emails when using a default mailer Mail.
-		if (
-			$this->mailer !== \WPMailSMTP\Options::init()->get( 'mail', 'mailer' )
-		) {
-			return true;
-		}
-
-		return false;
+		return true;
 	}
 }